+++ setup.py
@@ -27,7 +27,7 @@ base_packages = [
 dev_packages = [
     "graphviz>=0.10.1",
     "matplotlib>=3.0.2",
-    "mypy>=0.990",
+    "mypy>=1.11.1",
     "pre-commit>=2.20.0",
     "pytest>=7.2.0",
     "pytest-cov>=4.0.0",
